web-auto-agent/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
boto3==1.36.16
botocore==1.36.16
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
duckduckgo_search==7.3.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==25.1.24
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jmespath==1.0.1
keras==3.8.0
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
multidict==6.1.0
namex==0.0.8
numpy==2.0.2
openai==0.27.10
openai-function-call==0.0.5
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.48.0
pluggy==1.5.0
primp==0.12.0
propcache==0.2.1
protobuf==5.29.3
pycparser==2.22
pydantic==1.10.21
pyee==12.0.0
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3transfer==0.11.2
safetensors==0.5.2
selenium==4.28.1
six==1.17.0
smol-dev==0.0.4
smolagents==1.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
